+++ R/fscores.R
@@ -61,6 +61,12 @@
 #' \itemize{
 #'     \item \code{"EAP"} for the expected a-posteriori (default). For models fit using
 #'       \code{\link{mdirt}} this will return the posterior classification probabilities
+#'     \item \code{"EAP_general"} for the expected a-posteriori estimates using dimension reduction
+#'       to obtain estimates of the general factor(s) only (Gibbons, et al., 2007). This is generally
+#'       more accurate than the \code{'EAP'} approach as the dimensions to integrate are reduced to the
+#'       number of general factors plus one, and therefore higher number of quadrature nodes can be used
+#'       (see \code{quadpts} for the defaults. For a bifactor model, for example,
+#'       this method requires exactly 2 dimensions). Model must have been fit using \code{\link{bfactor}}
 #'     \item \code{"MAP"} for the maximum a-posteriori (i.e, Bayes modal)
 #'     \item \code{"ML"} for maximum likelihood
 #'     \item \code{"WLE"} or \code{"WML"} for weighted maximum-likelihood estimation
@@ -145,7 +151,7 @@
 #' factor models with applications in test scoring, scale alignment, and model fit testing.
 #' \emph{Psychometrika, 80}(2), 535–559.
 #'
-#' Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
+#' Chalmers, R. P. (2012). mirt: A Multidimensional Item Response Theory
 #' Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
 #' \doi{10.18637/jss.v048.i06}
 #'
@@ -154,6 +160,9 @@
 #' 1-39. \doi{10.18637/jss.v071.i05}
 #'
 #' Embretson, S. E. & Reise, S. P. (2000). Item Response Theory for Psychologists. Erlbaum.
+#'
+#' Gibbons. R. D., Bock, R. D., Hedeker, D., et al. (2007). Full-Information item bifactor
+#' analysis of graded response data. \emph{Applied Psychological Measurement, 31}(1), 4-19.
 #'
 #' Thissen, D., Pommerich, M., Billeaud, K., & Williams, V. S. L. (1995).
 #' Item Response Theory for Scores on Tests Including Polytomous Items with Ordered Responses.
@@ -303,6 +312,8 @@ fscores <- function(object, method = "EAP", full.scores = TRUE, rotate = 'oblimi
         if(QMC && is.null(quadpts)) quadpts <- 5000
         nfact <- object@Model$nfact
         if(method == 'EAPsum_2.0') nfact <- 2
+        if(method == 'EAP_general')
+            nfact <- nfact - attr(object@Model$model,"nspec") + 1
         if(is.null(quadpts))
             quadpts <- switch(as.character(nfact),
                               '1'=121, '2'=61, '3'=31, '4'=19, '5'=11, '6'=7, 5)