+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changes in mirt 1.47
 
+- `fscores()` also gains `method = 'EAP_general'` to perform the dimension reduction
+  EAP estimates for bifactor/two-tier general factor predictions as 
+  described in Gibbons et al. (2007)
+
 - `fscores()` gains `method = 'EAPsum_2.0'` to obtain the 2.0 version of the 
   Lord-Wingersky Algorithm described in Cai (2015)